requirements.txt
click==6.2
pocket-api
requests==2.9.1
